Move ORDER_BY, LIMIT and OFFSET parsing functions
